requirements.txt
async-timeout==5.0.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorzero==2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gpiozero==2.0.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
requests==2.32.3
RPi.GPIO==0.7.1
simple-websocket==1.1.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
